feat: add tablemode test
